Migrate everything to Fedora 42, and use Kickstart `liveimg`'s option
This commit do two important things. The first one is to use Fedora 42 instead of 41.

The second one is using Kickstart `liveimg`'s option. This allow Anaconda to easilly install the same apps than the LiveCD image. This avoid generating the repo and all that process.

Some other fixs here and there.